further moved HyperTransformerPre to a new file
